// src/api/generated/model/questionRef.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { Polarity } from './polarity';
import type { SignalCategory } from './signalCategory';
import type { Weight } from './weight';

export interface QuestionRef {
  id: string;
  key: string;
  text: string;
  category: SignalCategory;
  polarity: Polarity;
  weight: Weight;
}